Address agent hook review feedback
