+++ Sources/TerminalNotificationStore.swift
@@ -1,5 +1,4 @@
 import AppKit
-import Darwin
 import Foundation
 import UserNotifications
 import Bonsplit
@@ -660,640 +659,6 @@ enum TerminalNotificationAction: Hashable {
     case agentHookSetup(agentName: String)
 }
 
-struct AgentHookIntegration: Identifiable, Hashable, Sendable {
-    let name: String
-    let displayName: String
-    let commandNames: [String]
-    let configDir: String?
-    let configFile: String?
-    let configDirEnvOverride: String?
-    let hookMarkers: [String]
-    let currentMarkers: [String]
-    let isClaudeWrapper: Bool
-
-    var id: String { name }
-
-    var installCommand: String {
-        if isClaudeWrapper {
-            return "cmux settings open --section automation"
-        }
-        return "cmux hooks \(name) install"
-    }
-}
-
-enum AgentHookIntegrationStatus: Equatable {
-    case enabled
-    case disabled
-    case installed(path: String)
-    case updateAvailable(path: String)
-    case notInstalled(path: String?)
-    case unreadable(path: String)
-
-    var isActive: Bool {
-        switch self {
-        case .enabled, .installed:
-            return true
-        case .disabled, .updateAvailable, .notInstalled, .unreadable:
-            return false
-        }
-    }
-
-    var isUpdateAvailable: Bool {
-        if case .updateAvailable = self {
-            return true
-        }
-        return false
-    }
-}
-
-struct AgentHookInstallResult {
-    let succeeded: Bool
-    let message: String
-}
-
-struct AgentHookDiffResult {
-    let succeeded: Bool
-    let message: String
-    let diff: String
-}
-
-enum AgentHookIntegrationSettings {
-    static let promptEnabledKey = "agentHookSetupPromptEnabled"
-    static let defaultPromptEnabled = true
-    static let statusDidChangeNotification = Notification.Name("cmux.agentHookIntegration.statusDidChange")
-
-    private static let promptCooldown: TimeInterval = 24 * 60 * 60
-    private static let configFileWatcher = ConfigFileWatcher()
-
-    static let allAgents: [AgentHookIntegration] = [
-        AgentHookIntegration(
-            name: "claude",
-            displayName: "Claude Code",
-            commandNames: ["claude"],
-            configDir: nil,
-            configFile: nil,
-            configDirEnvOverride: nil,
-            hookMarkers: [],
-            currentMarkers: [],
-            isClaudeWrapper: true
-        ),
-        AgentHookIntegration(
-            name: "codex",
-            displayName: "Codex",
-            commandNames: ["codex"],
-            configDir: ".codex",
-            configFile: "hooks.json",
-            configDirEnvOverride: "CODEX_HOME",
-            hookMarkers: ["cmux hooks codex", "cmux codex-hook"],
-            currentMarkers: ["CMUX_AGENT_HOOK_VERSION=1"],
-            isClaudeWrapper: false
-        ),
-        AgentHookIntegration(
-            name: "opencode",
-            displayName: "OpenCode",
-            commandNames: ["opencode", "open-code"],
-            configDir: ".config/opencode",
-            configFile: "plugins/cmux-session.js",
-            configDirEnvOverride: "OPENCODE_CONFIG_DIR",
-            hookMarkers: ["cmux-opencode-session-plugin-marker", "cmux hooks opencode"],
-            currentMarkers: ["cmux-opencode-session-plugin-marker v1"],
-            isClaudeWrapper: false
-        ),
-        AgentHookIntegration(
-            name: "cursor",
-            displayName: "Cursor",
-            commandNames: ["cursor"],
-            configDir: ".cursor",
-            configFile: "hooks.json",
-            configDirEnvOverride: nil,
-            hookMarkers: ["cmux hooks cursor"],
-            currentMarkers: ["CMUX_AGENT_HOOK_VERSION=1"],
-            isClaudeWrapper: false
-        ),
-        AgentHookIntegration(
-            name: "gemini",
-            displayName: "Gemini",
-            commandNames: ["gemini"],
-            configDir: ".gemini",
-            configFile: "settings.json",
-            configDirEnvOverride: nil,
-            hookMarkers: ["cmux hooks gemini"],
-            currentMarkers: ["CMUX_AGENT_HOOK_VERSION=1"],
-            isClaudeWrapper: false
-        ),
-        AgentHookIntegration(
-            name: "copilot",
-            displayName: "Copilot",
-            commandNames: ["copilot"],
-            configDir: ".copilot",
-            configFile: "config.json",
-            configDirEnvOverride: nil,
-            hookMarkers: ["cmux hooks copilot"],
-            currentMarkers: ["CMUX_AGENT_HOOK_VERSION=1"],
-            isClaudeWrapper: false
-        ),
-        AgentHookIntegration(
-            name: "codebuddy",
-            displayName: "CodeBuddy",
-            commandNames: ["codebuddy"],
-            configDir: ".codebuddy",
-            configFile: "settings.json",
-            configDirEnvOverride: nil,
-            hookMarkers: ["cmux hooks codebuddy"],
-            currentMarkers: ["CMUX_AGENT_HOOK_VERSION=1"],
-            isClaudeWrapper: false
-        ),
-        AgentHookIntegration(
-            name: "factory",
-            displayName: "Factory",
-            commandNames: ["factory"],
-            configDir: ".factory",
-            configFile: "settings.json",
-            configDirEnvOverride: nil,
-            hookMarkers: ["cmux hooks factory"],
-            currentMarkers: ["CMUX_AGENT_HOOK_VERSION=1"],
-            isClaudeWrapper: false
-        ),
-        AgentHookIntegration(
-            name: "qoder",
-            displayName: "Qoder",
-            commandNames: ["qoder"],
-            configDir: ".qoder",
-            configFile: "settings.json",
-            configDirEnvOverride: nil,
-            hookMarkers: ["cmux hooks qoder"],
-            currentMarkers: ["CMUX_AGENT_HOOK_VERSION=1"],
-            isClaudeWrapper: false
-        ),
-    ]
-
-    static func promptEnabled(defaults: UserDefaults = .standard) -> Bool {
-        if defaults.object(forKey: promptEnabledKey) == nil {
-            return defaultPromptEnabled
-        }
-        return defaults.bool(forKey: promptEnabledKey)
-    }
-
-    static func setPromptEnabled(_ enabled: Bool, defaults: UserDefaults = .standard) {
-        defaults.set(enabled, forKey: promptEnabledKey)
-        NotificationCenter.default.post(name: statusDidChangeNotification, object: nil)
-    }
-
-    static func agent(named name: String) -> AgentHookIntegration? {
-        let normalized = name.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        return allAgents.first { agent in
-            agent.name == normalized || agent.commandNames.contains(normalized)
-        }
-    }
-
-    static func status(for agent: AgentHookIntegration, defaults: UserDefaults = .standard) -> AgentHookIntegrationStatus {
-        configFileWatcher.startIfNeeded()
-
-        if agent.isClaudeWrapper {
-            return ClaudeCodeIntegrationSettings.hooksEnabled(defaults: defaults) ? .enabled : .disabled
-        }
-
-        guard let path = configFilePath(for: agent) else {
-            return .notInstalled(path: nil)
-        }
-
-        guard FileManager.default.fileExists(atPath: path) else {
-            return .notInstalled(path: path)
-        }
-        guard let contents = try? String(contentsOfFile: path, encoding: .utf8) else {
-            return .unreadable(path: path)
-        }
-        if agent.currentMarkers.contains(where: { contents.contains($0) }) {
-            return .installed(path: path)
-        }
-        if agent.hookMarkers.contains(where: { contents.contains($0) }) {
-            return .updateAvailable(path: path)
-        }
-        return .notInstalled(path: path)
-    }
-
-    static func statusLabel(for status: AgentHookIntegrationStatus) -> String {
-        switch status {
-        case .enabled:
-            return String(localized: "settings.automation.agentHooks.status.enabled", defaultValue: "Enabled")
-        case .disabled:
-            return String(localized: "settings.automation.agentHooks.status.disabled", defaultValue: "Disabled")
-        case .installed:
-            return String(localized: "settings.automation.agentHooks.status.installed", defaultValue: "Installed")
-        case .updateAvailable:
-            return String(localized: "settings.automation.agentHooks.status.updateAvailable", defaultValue: "Update available")
-        case .notInstalled:
-            return String(localized: "settings.automation.agentHooks.status.notInstalled", defaultValue: "Not installed")
-        case .unreadable:
-            return String(localized: "settings.automation.agentHooks.status.unknown", defaultValue: "Unknown")
-        }
-    }
-
-    static func statusSubtitle(for agent: AgentHookIntegration, status: AgentHookIntegrationStatus) -> String {
-        switch status {
-        case .enabled:
-            return String(localized: "settings.automation.agentHooks.status.claudeEnabled", defaultValue: "cmux wraps the claude command in cmux terminals.")
-        case .disabled:
-            return String(localized: "settings.automation.agentHooks.status.claudeDisabled", defaultValue: "Claude Code runs without cmux hooks.")
-        case .installed(let path):
-            return String(localized: "settings.automation.agentHooks.status.installedAt", defaultValue: "cmux hooks found in \(path).")
-        case .updateAvailable:
-            return String(localized: "settings.automation.agentHooks.status.updateAvailable.subtitle", defaultValue: "cmux hooks are installed, but this app has a newer hook script.")
-        case .notInstalled:
-            return String(localized: "settings.automation.agentHooks.status.notInstalled.subtitle", defaultValue: "No cmux hooks found.")
-        case .unreadable(let path):
-            return String(localized: "settings.automation.agentHooks.status.unreadable", defaultValue: "Could not read \(path).")
-        }
-    }
-
-    @MainActor
-    static func showSetupPromptIfNeeded(agentName: String, tabId: UUID, surfaceId: UUID?) {
-        guard promptEnabled(),
-              let agent = agent(named: agentName) else {
-            return
-        }
-        let currentStatus = status(for: agent)
-        guard !currentStatus.isActive else {
-            return
-        }
-        guard shouldShowPrompt(for: agent, status: currentStatus) else {
-            return
-        }
-        guard !TerminalNotificationStore.shared.hasAgentHookSetupNotification(for: agent.name) else {
-            return
-        }
-
-        TerminalNotificationStore.shared.addNotification(
-            tabId: tabId,
-            surfaceId: surfaceId,
-            title: currentStatus.isUpdateAvailable
-                ? String(localized: "agentHooks.nudge.updateTitle", defaultValue: "Update \(agent.displayName) hooks")
-                : String(localized: "agentHooks.nudge.title", defaultValue: "Install \(agent.displayName) hooks"),
-            subtitle: String(localized: "agentHooks.nudge.subtitle", defaultValue: "Notifications and session restore"),
-            body: currentStatus.isUpdateAvailable
-                ? String(localized: "agentHooks.nudge.updateBody", defaultValue: "cmux has a newer hook script for notifications and session restore.")
-                : String(localized: "agentHooks.nudge.body", defaultValue: "Hooks let cmux show agent notifications and restore sessions after cmux restarts."),
-            action: .agentHookSetup(agentName: agent.name)
-        )
-        AppDelegate.shared?.toggleNotificationsPopover(animated: true)
-    }
-
-    static func snoozePrompt(agentName: String, defaults: UserDefaults = .standard) {
-        guard let agent = agent(named: agentName) else { return }
-        let currentStatus = status(for: agent, defaults: defaults)
-        markPromptSnoozed(for: agent, status: currentStatus, defaults: defaults)
-    }
-
-    static func installHooks(for agent: AgentHookIntegration, completion: @escaping (AgentHookInstallResult) -> Void) {
-        if agent.isClaudeWrapper {
-            UserDefaults.standard.set(true, forKey: ClaudeCodeIntegrationSettings.hooksEnabledKey)
-            NotificationCenter.default.post(name: statusDidChangeNotification, object: nil)
-            completion(AgentHookInstallResult(
-                succeeded: true,
-                message: String(localized: "settings.automation.agentHooks.status.claudeEnabled", defaultValue: "cmux wraps the claude command in cmux terminals.")
-            ))
-            return
-        }
-
-        let launch = hookInstallLaunch(for: agent)
-        DispatchQueue.global(qos: .userInitiated).async {
-            let result = runInstallCommand(
-                executableURL: launch.executableURL,
-                arguments: launch.arguments,
-                environment: nil,
-                fallbackCommand: agent.installCommand
-            )
-            DispatchQueue.main.async {
-                configFileWatcher.refreshWatchedPaths()
-                NotificationCenter.default.post(name: statusDidChangeNotification, object: nil)
-                completion(result)
-            }
-        }
-    }
-
-    static func diffHooks(for agent: AgentHookIntegration, completion: @escaping (AgentHookDiffResult) -> Void) {
-        if agent.isClaudeWrapper {
-            completion(AgentHookDiffResult(
-                succeeded: true,
-                message: "",
-                diff: String(localized: "agentHooks.diff.claude", defaultValue: "Claude Code uses the cmux wrapper in cmux terminals. No config file changes are needed.")
-            ))
-            return
-        }
-
-        DispatchQueue.global(qos: .userInitiated).async {
-            let result = buildHookDiff(for: agent)
-            DispatchQueue.main.async {
-                completion(result)
-            }
-        }
-    }
-
-    private static func configDirectoryPath(for agent: AgentHookIntegration) -> String? {
-        guard let configDir = agent.configDir else {
-            return nil
-        }
-        if let envKey = agent.configDirEnvOverride,
-           let envValue = ProcessInfo.processInfo.environment[envKey],
-           !envValue.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-            return NSString(string: envValue).expandingTildeInPath
-        }
-        return NSString(string: "~/\(configDir)").expandingTildeInPath
-    }
-
-    private static func configFilePath(for agent: AgentHookIntegration) -> String? {
-        guard let directory = configDirectoryPath(for: agent),
-              let configFile = agent.configFile else {
-            return nil
-        }
-        return (directory as NSString).appendingPathComponent(configFile)
-    }
-
-    private static func shouldShowPrompt(
-        for agent: AgentHookIntegration,
-        status: AgentHookIntegrationStatus,
-        defaults: UserDefaults = .standard
-    ) -> Bool {
-        let key = lastPromptKey(for: agent, status: status)
-        let lastSnoozed = defaults.double(forKey: key)
-        guard lastSnoozed > 0 else { return true }
-        return Date().timeIntervalSince1970 - lastSnoozed >= promptCooldown
-    }
-
-    private static func markPromptSnoozed(
-        for agent: AgentHookIntegration,
-        status: AgentHookIntegrationStatus,
-        defaults: UserDefaults = .standard
-    ) {
-        defaults.set(Date().timeIntervalSince1970, forKey: lastPromptKey(for: agent, status: status))
-    }
-
-    private static func lastPromptKey(for agent: AgentHookIntegration, status: AgentHookIntegrationStatus) -> String {
-        let kind = status.isUpdateAvailable ? "update" : "install"
-        return "agentHookSetupPromptSnoozedAt.\(agent.name).\(kind)"
-    }
-
-    private static func hookInstallLaunch(for agent: AgentHookIntegration) -> (executableURL: URL, arguments: [String]) {
-        if let bundledCLIURL = Bundle.main.resourceURL?.appendingPathComponent("bin/cmux"),
-           FileManager.default.isExecutableFile(atPath: bundledCLIURL.path) {
-            return (bundledCLIURL, ["hooks", agent.name, "install", "--yes"])
-        }
-        return (URL(fileURLWithPath: "/usr/bin/env"), ["cmux", "hooks", agent.name, "install", "--yes"])
-    }
-
-    private static func buildHookDiff(for agent: AgentHookIntegration) -> AgentHookDiffResult {
-        let fm = FileManager.default
-        let tempHome = URL(fileURLWithPath: NSTemporaryDirectory(), isDirectory: true)
-            .appendingPathComponent("cmux-agent-hook-diff-\(UUID().uuidString)", isDirectory: true)
-
-        do {
-            try fm.createDirectory(at: tempHome, withIntermediateDirectories: true)
-            defer { try? fm.removeItem(at: tempHome) }
-
-            guard let configDir = agent.configDir else {
-                return AgentHookDiffResult(
-                    succeeded: false,
-                    message: String(localized: "agentHooks.diff.failed", defaultValue: "Could not prepare hook diff."),
-                    diff: ""
-                )
-            }
-
-            let originalConfigDir = expandedHomePath(configDir)
-            let tempConfigDir = tempHome.appendingPathComponent(configDir, isDirectory: true)
-            try fm.createDirectory(at: tempConfigDir.deletingLastPathComponent(), withIntermediateDirectories: true)
-            if fm.fileExists(atPath: originalConfigDir.path) {
-                try fm.copyItem(at: originalConfigDir, to: tempConfigDir)
-            } else {
-                try fm.createDirectory(at: tempConfigDir, withIntermediateDirectories: true)
-            }
-
-            var environment = ProcessInfo.processInfo.environment
-            environment["HOME"] = tempHome.path
-            if let envKey = agent.configDirEnvOverride {
-                environment[envKey] = tempConfigDir.path
-            }
-
-            let launch = hookInstallLaunch(for: agent)
-            let installResult = runInstallCommand(
-                executableURL: launch.executableURL,
-                arguments: launch.arguments,
-                environment: environment,
-                fallbackCommand: agent.installCommand
-            )
-            guard installResult.succeeded else {
-                return AgentHookDiffResult(succeeded: false, message: installResult.message, diff: "")
-            }
-
-            let relativePaths = diffRelativePaths(for: agent)
-            let diffs = relativePaths.compactMap { relativePath in
-                let oldURL = URL(fileURLWithPath: NSString(string: "~/\(relativePath)").expandingTildeInPath)
-                let newURL = tempHome.appendingPathComponent(relativePath)
-                return unifiedDiff(relativePath: relativePath, oldURL: oldURL, newURL: newURL)
-            }
-            let diff = diffs.joined(separator: "\n")
-            if diff.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-                return AgentHookDiffResult(
-                    succeeded: true,
-                    message: String(localized: "agentHooks.diff.noChanges", defaultValue: "No file changes needed."),
-                    diff: String(localized: "agentHooks.diff.noChanges", defaultValue: "No file changes needed.")
-                )
-            }
-            return AgentHookDiffResult(succeeded: true, message: "", diff: diff)
-        } catch {
-            return AgentHookDiffResult(
-                succeeded: false,
-                message: String(localized: "agentHooks.diff.failed", defaultValue: "Could not prepare hook diff."),
-                diff: ""
-            )
-        }
-    }
-
-    private static func expandedHomePath(_ relativePath: String) -> URL {
-        URL(fileURLWithPath: NSString(string: "~/\(relativePath)").expandingTildeInPath)
-    }
-
-    private static func diffRelativePaths(for agent: AgentHookIntegration) -> [String] {
-        guard let configDir = agent.configDir,
-              let configFile = agent.configFile else {
-            return []
-        }
-        var paths = ["\(configDir)/\(configFile)"]
-        if agent.name == "codex" {
-            paths.append("\(configDir)/config.toml")
-        }
-        return paths
-    }
-
-    private static func unifiedDiff(relativePath: String, oldURL: URL, newURL: URL) -> String? {
-        let oldText = (try? String(contentsOf: oldURL, encoding: .utf8)) ?? ""
-        let newText = (try? String(contentsOf: newURL, encoding: .utf8)) ?? ""
-        guard oldText != newText else { return nil }
-
-        let oldLines = oldText.split(separator: "\n", omittingEmptySubsequences: false).map(String.init)
-        let newLines = newText.split(separator: "\n", omittingEmptySubsequences: false).map(String.init)
-        var lines = [
-            "--- ~/\(relativePath)",
-            "+++ ~/\(relativePath)",
-            "@@",
-        ]
-        lines.append(contentsOf: oldLines.map { "-\($0)" })
-        lines.append(contentsOf: newLines.map { "+\($0)" })
-        return lines.joined(separator: "\n")
-    }
-
-    private static func watchedConfigFilePaths() -> Set<String> {
-        var paths: Set<String> = []
-        for agent in allAgents where !agent.isClaudeWrapper {
-            if let configFilePath = configFilePath(for: agent) {
-                paths.insert(configFilePath)
-            }
-            if agent.name == "codex",
-               let configDirectoryPath = configDirectoryPath(for: agent) {
-                paths.insert((configDirectoryPath as NSString).appendingPathComponent("config.toml"))
-            }
-        }
-        return paths
-    }
-
-    private static func watchedConfigPaths() -> Set<String> {
-        let fm = FileManager.default
-        var paths: Set<String> = []
-        for filePath in watchedConfigFilePaths() {
-            let fileURL = URL(fileURLWithPath: filePath).standardizedFileURL
-            if fm.fileExists(atPath: fileURL.path) {
-                paths.insert(fileURL.path)
-            }
-            if let parentURL = nearestExistingAncestor(for: fileURL.deletingLastPathComponent()) {
-                paths.insert(parentURL.path)
-            }
-        }
-        return paths
-    }
-
-    private static func nearestExistingAncestor(for url: URL) -> URL? {
-        let fm = FileManager.default
-        var current = url.standardizedFileURL
-        while true {
-            if fm.fileExists(atPath: current.path) {
-                return current
-            }
-            let parent = current.deletingLastPathComponent().standardizedFileURL
-            if parent.path == current.path {
-                return nil
-            }
-            current = parent
-        }
-    }
-
-    private final class ConfigFileWatcher {
-        private let queue = DispatchQueue(label: "com.cmuxterm.agent-hook-config-watcher", qos: .utility)
-        private var isStarted = false
-        private var watchedPaths: Set<String> = []
-        private var sources: [String: DispatchSourceFileSystemObject] = [:]
-
-        func startIfNeeded() {
-            queue.async { [weak self] in
-                guard let self, !isStarted else { return }
-                isStarted = true
-                rebuildWatchedPaths()
-            }
-        }
-
-        func refreshWatchedPaths() {
-            queue.async { [weak self] in
-                guard let self, isStarted else { return }
-                rebuildWatchedPaths()
-            }
-        }
-
-        private func rebuildWatchedPaths() {
-            let nextPaths = AgentHookIntegrationSettings.watchedConfigPaths()
-            for path in watchedPaths.subtracting(nextPaths) {
-                sources.removeValue(forKey: path)?.cancel()
-            }
-            for path in nextPaths.subtracting(watchedPaths) {
-                startWatching(path: path)
-            }
-            watchedPaths = Set(sources.keys)
-        }
-
-        private func startWatching(path: String) {
-            let descriptor = open(path, O_EVTONLY)
-            guard descriptor >= 0 else { return }
-            let source = DispatchSource.makeFileSystemObjectSource(
-                fileDescriptor: descriptor,
-                eventMask: [.write, .delete, .rename, .attrib, .extend, .link, .revoke],
-                queue: queue
-            )
-            source.setEventHandler { [weak self] in
-                self?.handleChange()
-            }
-            source.setCancelHandler {
-                close(descriptor)
-            }
-            sources[path] = source
-            source.resume()
-        }
-
-        private func handleChange() {
-            rebuildWatchedPaths()
-            DispatchQueue.main.async {
-                NotificationCenter.default.post(name: AgentHookIntegrationSettings.statusDidChangeNotification, object: nil)
-            }
-        }
-    }
-
-    private static func runInstallCommand(
-        executableURL: URL,
-        arguments: [String],
-        environment: [String: String]?,
-        fallbackCommand: String
-    ) -> AgentHookInstallResult {
-        let process = Process()
-        process.executableURL = executableURL
-        process.arguments = arguments
-        if let environment {
-            process.environment = environment
-        }
-        process.standardInput = FileHandle.nullDevice
-        let stdoutPipe = Pipe()
-        let stderrPipe = Pipe()
-        process.standardOutput = stdoutPipe
-        process.standardError = stderrPipe
-
-        do {
-            try process.run()
-            process.waitUntilExit()
-        } catch {
-            return AgentHookInstallResult(
-                succeeded: false,
-                message: String(localized: "agentHooks.prompt.installFailed", defaultValue: "Could not install hooks. Run \(fallbackCommand) in a terminal.")
-            )
-        }
-
-        let outputData = stdoutPipe.fileHandleForReading.readDataToEndOfFile()
-        let errorData = stderrPipe.fileHandleForReading.readDataToEndOfFile()
-        let output = String(data: outputData, encoding: .utf8)?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-        let errorOutput = String(data: errorData, encoding: .utf8)?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-
-        guard process.terminationStatus == 0 else {
-            let detail = errorOutput.isEmpty ? output : errorOutput
-            if detail.isEmpty {
-                return AgentHookInstallResult(
-                    succeeded: false,
-                    message: String(localized: "agentHooks.prompt.installFailed", defaultValue: "Could not install hooks. Run \(fallbackCommand) in a terminal.")
-                )
-            }
-            return AgentHookInstallResult(succeeded: false, message: detail)
-        }
-
-        return AgentHookInstallResult(
-            succeeded: true,
-            message: String(localized: "agentHooks.prompt.installSucceeded", defaultValue: "Hooks installed.")
-        )
-    }
-}
-
 struct TerminalNotification: Identifiable, Hashable {
     let id: UUID
     let tabId: UUID
@@ -1590,7 +955,9 @@ final class TerminalNotificationStore: ObservableObject {
         var updated = notifications
         var idsToClear: [String] = []
         updated.removeAll { existing in
-            guard existing.tabId == tabId, existing.surfaceId == surfaceId else { return false }
+            guard Self.shouldReplaceNotification(existing, tabId: tabId, surfaceId: surfaceId, action: action) else {
+                return false
+            }
             idsToClear.append(existing.id.uuidString)
             return true
         }
@@ -1638,6 +1005,23 @@ final class TerminalNotificationStore: ObservableObject {
             suppressedNotificationFeedbackHandler(self, notification)
         } else {
             notificationDeliveryHandler(self, notification)
+        }
+    }
+
+    private static func shouldReplaceNotification(
+        _ existing: TerminalNotification,
+        tabId: UUID,
+        surfaceId: UUID?,
+        action: TerminalNotificationAction?
+    ) -> Bool {
+        guard existing.tabId == tabId, existing.surfaceId == surfaceId else { return false }
+        switch (existing.action, action) {
+        case (.agentHookSetup(let existingAgent)?, .agentHookSetup(let newAgent)?):
+            return existingAgent == newAgent
+        case (nil, nil):
+            return true
+        default:
+            return false
         }
     }
